PgSQL to LDAP transition - Phase 1: Added LDAP support for member data.
This is the first major update for the PgSQL to LDAP transition. We will be
deprecating the members table in PostgreSQL and moving that information into
the LDAP directory.

With this commit, all newly created members will have their attributes added
to LDAP in addition to the members table in the database. The database is still
used as the canonical source of this information: the information in LDAP is
kept up-to-date but almost never referenced.